requirements.txt
# Python Requirements for AI Resume Tailoring System

# OpenAI API
openai>=1.12.0

# PDF Processing
pypdf>=3.0.0

# LinkedIn Job Automation
selenium>=4.16.0
requests>=2.31.0
webdriver-manager>=4.0.0

# Document Processing
python-docx>=1.1.0
